requirements.txt
PyQt6>=6.6
opencv-python>=4.9
mediapipe>=0.10
numpy>=1.26
PyYAML>=6.0
